Update to latest rails main
- Drop Rails 7.0 and 7.1 tests, add separate 7.2 tests
- Update behaviour tests from main
- Add instrumentation for increment and decrement
- Don't lock the gemfiles
