--- gemfiles/rails_7_2.gemfile.lock
@@ -1,236 +1,0 @@
-GIT
-  remote: https://github.com/rails/rails.git
-  revision: 2409401c87d512c6e505f3e00110558aa62a94f7
-  branch: 7-2-stable
-  specs:
-    actionpack (7.2.0)
-      actionview (= 7.2.0)
-      activesupport (= 7.2.0)
-      nokogiri (>= 1.8.5)
-      racc
-      rack (>= 2.2.4, < 3.2)
-      rack-session (>= 1.0.1)
-      rack-test (>= 0.6.3)
-      rails-dom-testing (~> 2.2)
-      rails-html-sanitizer (~> 1.6)
-      useragent (~> 0.16)
-    actionview (7.2.0)
-      activesupport (= 7.2.0)
-      builder (~> 3.1)
-      erubi (~> 1.11)
-      rails-dom-testing (~> 2.2)
-      rails-html-sanitizer (~> 1.6)
-    activejob (7.2.0)
-      activesupport (= 7.2.0)
-      globalid (>= 0.3.6)
-    activemodel (7.2.0)
-      activesupport (= 7.2.0)
-    activerecord (7.2.0)
-      activemodel (= 7.2.0)
-      activesupport (= 7.2.0)
-      timeout (>= 0.4.0)
-    activesupport (7.2.0)
-      base64
-      bigdecimal
-      concurrent-ruby (~> 1.0, >= 1.3.1)
-      connection_pool (>= 2.2.5)
-      drb
-      i18n (>= 1.6, < 2)
-      logger (>= 1.4.2)
-      minitest (>= 5.1)
-      securerandom (>= 0.3)
-      tzinfo (~> 2.0, >= 2.0.5)
-    railties (7.2.0)
-      actionpack (= 7.2.0)
-      activesupport (= 7.2.0)
-      irb (~> 1.13)
-      rackup (>= 1.0.0)
-      rake (>= 12.2)
-      thor (~> 1.0, >= 1.2.2)
-      zeitwerk (~> 2.6)
-
-PATH
-  remote: ..
-  specs:
-    solid_cache (1.0.1)
-      activejob (>= 7.2)
-      activerecord (>= 7.2)
-      railties (>= 7.2)
-
-GEM
-  remote: https://rubygems.org/
-  specs:
-    appraisal (2.5.0)
-      bundler
-      rake
-      thor (>= 0.14.0)
-    ast (2.4.2)
-    base64 (0.2.0)
-    bigdecimal (3.1.8)
-    builder (3.3.0)
-    concurrent-ruby (1.3.4)
-    connection_pool (2.4.1)
-    crass (1.0.6)
-    debug (1.9.2)
-      irb (~> 1.10)
-      reline (>= 0.3.8)
-    drb (2.2.1)
-    erubi (1.13.0)
-    globalid (1.2.1)
-      activesupport (>= 6.1)
-    i18n (1.14.5)
-      concurrent-ruby (~> 1.0)
-    io-console (0.7.2)
-    irb (1.14.0)
-      rdoc (>= 4.0.0)
-      reline (>= 0.4.2)
-    json (2.7.2)
-    language_server-protocol (3.17.0.3)
-    logger (1.6.0)
-    loofah (2.22.0)
-      crass (~> 1.0.2)
-      nokogiri (>= 1.12.0)
-    minitest (5.24.1)
-    mocha (2.4.5)
-      ruby2_keywords (>= 0.0.5)
-    msgpack (1.7.2)
-    mysql2 (0.5.6)
-    nokogiri (1.16.7-aarch64-linux)
-      racc (~> 1.4)
-    nokogiri (1.16.7-arm-linux)
-      racc (~> 1.4)
-    nokogiri (1.16.7-arm64-darwin)
-      racc (~> 1.4)
-    nokogiri (1.16.7-x86-linux)
-      racc (~> 1.4)
-    nokogiri (1.16.7-x86_64-darwin)
-      racc (~> 1.4)
-    nokogiri (1.16.7-x86_64-linux)
-      racc (~> 1.4)
-    parallel (1.26.2)
-    parser (3.3.4.2)
-      ast (~> 2.4.1)
-      racc
-    pg (1.5.7)
-    psych (5.1.2)
-      stringio
-    racc (1.8.1)
-    rack (3.1.7)
-    rack-session (2.0.0)
-      rack (>= 3.0.0)
-    rack-test (2.1.0)
-      rack (>= 1.3)
-    rackup (2.1.0)
-      rack (>= 3)
-      webrick (~> 1.8)
-    rails-dom-testing (2.2.0)
-      activesupport (>= 5.0.0)
-      minitest
-      nokogiri (>= 1.6)
-    rails-html-sanitizer (1.6.0)
-      loofah (~> 2.21)
-      nokogiri (~> 1.14)
-    rainbow (3.1.1)
-    rake (13.2.1)
-    rdoc (6.7.0)
-      psych (>= 4.0.0)
-    regexp_parser (2.9.2)
-    reline (0.5.9)
-      io-console (~> 0.5)
-    rexml (3.3.5)
-      strscan
-    rubocop (1.65.1)
-      json (~> 2.3)
-      language_server-protocol (>= 3.17.0)
-      parallel (~> 1.10)
-      parser (>= 3.3.0.2)
-      rainbow (>= 2.2.2, < 4.0)
-      regexp_parser (>= 2.4, < 3.0)
-      rexml (>= 3.2.5, < 4.0)
-      rubocop-ast (>= 1.31.1, < 2.0)
-      ruby-progressbar (~> 1.7)
-      unicode-display_width (>= 2.4.0, < 3.0)
-    rubocop-ast (1.32.0)
-      parser (>= 3.3.1.0)
-    rubocop-md (1.2.2)
-      rubocop (>= 1.0)
-    rubocop-minitest (0.35.1)
-      rubocop (>= 1.61, < 2.0)
-      rubocop-ast (>= 1.31.1, < 2.0)
-    rubocop-packaging (0.5.2)
-      rubocop (>= 1.33, < 2.0)
-    rubocop-performance (1.21.1)
-      rubocop (>= 1.48.1, < 2.0)
-      rubocop-ast (>= 1.31.1, < 2.0)
-    rubocop-rails (2.25.1)
-      activesupport (>= 4.2.0)
-      rack (>= 1.1)
-      rubocop (>= 1.33.0, < 2.0)
-      rubocop-ast (>= 1.31.1, < 2.0)
-    ruby-progressbar (1.13.0)
-    ruby2_keywords (0.0.5)
-    securerandom (0.3.1)
-    sprockets (4.2.1)
-      concurrent-ruby (~> 1.0)
-      rack (>= 2.2.4, < 4)
-    sprockets-rails (3.5.2)
-      actionpack (>= 6.1)
-      activesupport (>= 6.1)
-      sprockets (>= 3.0.0)
-    sqlite3 (2.0.3-aarch64-linux-gnu)
-    sqlite3 (2.0.3-aarch64-linux-musl)
-    sqlite3 (2.0.3-arm-linux-gnu)
-    sqlite3 (2.0.3-arm-linux-musl)
-    sqlite3 (2.0.3-arm64-darwin)
-    sqlite3 (2.0.3-x86-linux-gnu)
-    sqlite3 (2.0.3-x86-linux-musl)
-    sqlite3 (2.0.3-x86_64-darwin)
-    sqlite3 (2.0.3-x86_64-linux-gnu)
-    sqlite3 (2.0.3-x86_64-linux-musl)
-    stringio (3.1.1)
-    strscan (3.1.0)
-    thor (1.3.1)
-    timeout (0.4.1)
-    tzinfo (2.0.6)
-      concurrent-ruby (~> 1.0)
-    unicode-display_width (2.5.0)
-    useragent (0.16.10)
-    webrick (1.8.1)
-    zeitwerk (2.6.17)
-
-PLATFORMS
-  aarch64-linux
-  aarch64-linux-gnu
-  aarch64-linux-musl
-  arm-linux
-  arm-linux-gnu
-  arm-linux-musl
-  arm64-darwin
-  x86-linux
-  x86-linux-gnu
-  x86-linux-musl
-  x86_64-darwin
-  x86_64-linux
-  x86_64-linux-gnu
-  x86_64-linux-musl
-
-DEPENDENCIES
-  appraisal
-  debug
-  mocha
-  msgpack
-  mysql2
-  pg
-  railties!
-  rubocop (>= 1.25.1)
-  rubocop-md
-  rubocop-minitest
-  rubocop-packaging
-  rubocop-performance
-  rubocop-rails
-  solid_cache!
-  sprockets-rails
-  sqlite3
-
-BUNDLED WITH
-   2.5.9
