+++ test/unit/behaviors.rb
@@ -1,33 +1,21 @@
 # frozen_string_literal: true
 
-if Rails::VERSION::MAJOR == 7 && Rails::VERSION::MINOR == 0
-  require_relative "behaviors_rails_7_0/cache_delete_matched_behavior"
-  require_relative "behaviors_rails_7_0/cache_increment_decrement_behavior"
-  require_relative "behaviors_rails_7_0/cache_instrumentation_behavior"
-  require_relative "behaviors_rails_7_0/cache_store_behavior"
-  require_relative "behaviors_rails_7_0/cache_store_version_behavior"
-  require_relative "behaviors_rails_7_0/cache_store_coder_behavior"
-  require_relative "behaviors_rails_7_0/connection_pool_behavior"
-  require_relative "behaviors_rails_7_0/encoded_key_cache_behavior"
-  require_relative "behaviors_rails_7_0/failure_safety_behavior"
-  require_relative "behaviors_rails_7_0/failure_raising_behavior"
-  require_relative "behaviors_rails_7_0/local_cache_behavior"
-elsif Rails::VERSION::MAJOR == 7 && Rails::VERSION::MINOR == 1
-  require_relative "behaviors_rails_7_1/cache_delete_matched_behavior"
-  require_relative "behaviors_rails_7_1/cache_increment_decrement_behavior"
-  require_relative "behaviors_rails_7_1/cache_instrumentation_behavior"
-  require_relative "behaviors_rails_7_1/cache_logging_behavior"
-  require_relative "behaviors_rails_7_1/cache_store_behavior"
-  require_relative "behaviors_rails_7_1/cache_store_version_behavior"
-  require_relative "behaviors_rails_7_1/cache_store_coder_behavior"
-  require_relative "behaviors_rails_7_1/cache_store_compression_behavior"
-  require_relative "behaviors_rails_7_1/cache_store_format_version_behavior"
-  require_relative "behaviors_rails_7_1/cache_store_serializer_behavior"
-  require_relative "behaviors_rails_7_1/connection_pool_behavior"
-  require_relative "behaviors_rails_7_1/encoded_key_cache_behavior"
-  require_relative "behaviors_rails_7_1/failure_safety_behavior"
-  require_relative "behaviors_rails_7_1/failure_raising_behavior"
-  require_relative "behaviors_rails_7_1/local_cache_behavior"
+if Rails::VERSION::MAJOR == 7 && Rails::VERSION::MINOR == 2
+  require_relative "behaviors_rails_7_2/cache_delete_matched_behavior"
+  require_relative "behaviors_rails_7_2/cache_increment_decrement_behavior"
+  require_relative "behaviors_rails_7_2/cache_instrumentation_behavior"
+  require_relative "behaviors_rails_7_2/cache_logging_behavior"
+  require_relative "behaviors_rails_7_2/cache_store_behavior"
+  require_relative "behaviors_rails_7_2/cache_store_version_behavior"
+  require_relative "behaviors_rails_7_2/cache_store_coder_behavior"
+  require_relative "behaviors_rails_7_2/cache_store_compression_behavior"
+  require_relative "behaviors_rails_7_2/cache_store_format_version_behavior"
+  require_relative "behaviors_rails_7_2/cache_store_serializer_behavior"
+  require_relative "behaviors_rails_7_2/connection_pool_behavior"
+  require_relative "behaviors_rails_7_2/encoded_key_cache_behavior"
+  require_relative "behaviors_rails_7_2/failure_safety_behavior"
+  require_relative "behaviors_rails_7_2/failure_raising_behavior"
+  require_relative "behaviors_rails_7_2/local_cache_behavior"
 else
   require_relative "behaviors/cache_delete_matched_behavior"
   require_relative "behaviors/cache_increment_decrement_behavior"
