+++ test/unit/behaviors/cache_store_compression_behavior.rb
@@ -125,9 +125,7 @@ module CacheStoreCompressionBehavior
     LARGE_OBJECT = { data: LARGE_STRING }
 
     def with_format(format_version, &block)
-      ActiveSupport.deprecator.silence do
-        ActiveSupport::Cache.with(format_version: format_version, &block)
-      end
+      ActiveSupport::Cache.with(format_version: format_version, &block)
     end
 
     def assert_compress(value, **options)
