+++ test/unit/behaviors/cache_store_serializer_behavior.rb
@@ -44,11 +44,4 @@ module CacheStoreSerializerBehavior
       end
     end
   end
-
-  private
-    def with_format(format_version, &block)
-      ActiveSupport.deprecator.silence do
-        ActiveSupport::Cache.with(format_version: format_version, &block)
-      end
-    end
 end
