+++ test/unit/behaviors_rails_7_2/cache_delete_matched_behavior.rb
