+++ test/unit/behaviors_rails_7_2/cache_increment_decrement_behavior.rb
@@ -26,6 +26,22 @@ module CacheIncrementDecrementBehavior
     assert_equal 1, @cache.read(key, raw: true).to_i
 
     missing = @cache.decrement(SecureRandom.alphanumeric)
-    assert(missing.nil? || missing == -1)
+    assert_equal -1, missing
+    missing = @cache.decrement(SecureRandom.alphanumeric, 100)
+    assert_equal -100, missing
+  end
+
+  def test_ttl_isnt_updated
+    key = SecureRandom.uuid
+
+    assert_equal 1, @cache.increment(key, 1, expires_in: 1)
+    assert_equal 2, @cache.increment(key, 1, expires_in: 5000)
+
+    # having to sleep two seconds in a test is bad, but we're testing
+    # a wide range of backends with different TTL mecanisms, most without
+    # subsecond granularity, so this is the only reliable way.
+    sleep 2
+
+    assert_nil @cache.read(key, raw: true)
   end
 end
