+++ test/unit/behaviors_rails_7_2/cache_instrumentation_behavior.rb
@@ -34,6 +34,21 @@ module CacheInstrumentationBehavior
     assert_equal @cache.class.name, events[0].payload[:store]
   end
 
+  def test_fetch_multi_instrumentation_order_of_operations
+    skip if Rails::VERSION::MAJOR == 7 && Rails::VERSION::MINOR < 2
+    operations = []
+    callback = ->(name, *) { operations << name }
+
+    key_1 = SecureRandom.uuid
+    key_2 = SecureRandom.uuid
+
+    ActiveSupport::Notifications.subscribed(callback, /^cache_(read_multi|write_multi)\.active_support$/) do
+      @cache.fetch_multi(key_1, key_2) { |key| key * 2 }
+    end
+
+    assert_equal %w[ cache_read_multi.active_support cache_write_multi.active_support ], operations
+  end
+
   def test_read_multi_instrumentation
     key_1 = SecureRandom.uuid
     @cache.write(key_1, SecureRandom.alphanumeric)
@@ -55,9 +70,7 @@ module CacheInstrumentationBehavior
       event_name = "cache_#{method}.active_support"
 
       [].tap do |events|
-        ActiveSupport::Notifications.subscribe event_name do |*args|
-          events << ActiveSupport::Notifications::Event.new(*args)
-        end
+        ActiveSupport::Notifications.subscribe(event_name) { |event| events << event }
         yield
       end
     ensure
