+++ test/unit/behaviors_rails_7_2/cache_store_coder_behavior.rb
