+++ test/unit/behaviors_rails_7_2/cache_store_format_version_behavior.rb
@@ -43,9 +43,9 @@ module CacheStoreFormatVersionBehavior
           lookup_store.send(:serialize_entry, ActiveSupport::Cache::Entry.new(["value"] * 100))
         end
 
-        skip if !serialized.is_a?(String)
-
-        assert_operator serialized, :start_with?, uncompressed_signature
+        if serialized.is_a?(String)
+          assert_operator serialized, :start_with?, uncompressed_signature
+        end
       end
 
       test "format version #{format_version.inspect} uses correct signature for compressed entries" do
@@ -53,9 +53,9 @@ module CacheStoreFormatVersionBehavior
           lookup_store.send(:serialize_entry, ActiveSupport::Cache::Entry.new(["value"] * 100), compress_threshold: 1)
         end
 
-        skip if !serialized.is_a?(String)
-
-        assert_operator serialized, :start_with?, compressed_signature
+        if serialized.is_a?(String)
+          assert_operator serialized, :start_with?, compressed_signature
+        end
       end
 
       test "Marshal undefined class/module deserialization error with #{format_version} format" do
