+++ test/unit/behaviors_rails_7_2/connection_pool_behavior.rb
