+++ test/unit/behaviors_rails_7_2/encoded_key_cache_behavior.rb
