+++ test/unit/behaviors_rails_7_2/failure_safety_behavior.rb
@@ -49,7 +49,7 @@ module FailureSafetyBehavior
     end
   end
 
-  def test_write_failure_returns_false
+  def test_write_failure_returns_nil
     key = SecureRandom.uuid
     emulating_unavailability do |cache|
       assert_nil cache.write(key, SecureRandom.alphanumeric)
@@ -88,6 +88,19 @@ module FailureSafetyBehavior
 
     emulating_unavailability do |cache|
       assert_equal false, cache.delete(key)
+    end
+  end
+
+  def test_delete_multi_failure_returns_zero
+    key = SecureRandom.uuid
+    other_key = SecureRandom.uuid
+    @cache.write_multi(
+      key => SecureRandom.alphanumeric,
+      other_key => SecureRandom.alphanumeric
+    )
+
+    emulating_unavailability do |cache|
+      assert_equal 0, cache.delete_multi([key, other_key])
     end
   end
